Add SVG mixin required by o-buttons
Unfortunately, a minor release to o-buttons (v3.1.0) introduced a dependency on
o-ft-icons@^3.3.0, which is now causing conflicts for build service users at
bundles that were working previously.

This patch adds the SVG mixin that o-buttons needs so that we can widen the
semver range for o-ft-icons, which will prevent the conflict.